test: extract keyboard interaction mode tests
