+++ packages/grid/test/keyboard-interaction-mode-lit.test.js
@@ -1,0 +1,3 @@
+import '../theme/lumo/lit-all-imports.js';
+import '../src/lit-all-imports.js';
+import './keyboard-navigation.common.js';
