+++ packages/grid/test/keyboard-interaction-mode-polymer.test.js
@@ -1,0 +1,2 @@
+import '../all-imports.js';
+import './keyboard-interaction-mode.common.js';
